monitor/gcp_usage_report_to_slack/requirement.txt
google-cloud-bigquery
pandas
db-dtypes
datetime
slack_sdk
requests